requirements.txt
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.0.0
mlflow>=2.9.0
timm>=0.9.12
PyYAML>=6.0
onnx>=1.15.0
onnxruntime>=1.16.0
gradio>=4.0.0
huggingface-hub>=0.20.0
matplotlib>=3.7.0
numpy>=1.24.0
pytest>=7.4.0
